+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+## [v3.4.0-beta.4]
+  - 📅 2025-03-20
   - 🌟 [@mixte/snippets] 为 `Ant Design X` 新增了一个 `useReactFunction` 工具方法, 用于在 Vue 环境中执行 React 函数并捕获其返回值 *
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `useXAgent` 和 `useXChat` 导出 *
   - 🐞 [@mixte/snippets] 修复 `Ant Design X` 部分组件无法使用的问题 *
@@ -99,7 +102,9 @@ CHANGELOG 图标规范
 <hr>
 <br>
 
-[Unreleased]: https://github.com/MoomFE/mixte/compare/v3.4.0-beta.2...HEAD
+[Unreleased]: https://github.com/MoomFE/mixte/compare/v3.4.0-beta.4...HEAD
+[v3.4.0-beta.4]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.4
+[v3.4.0-beta.3]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.3
 [v3.4.0-beta.2]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.2
 [v3.4.0-beta.1]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.1
 [v3.3.0]: https://github.com/MoomFE/mixte/releases/tag/v3.3.0